+++ apps/server/src/core/error/interface/error-type.interface.ts
@@ -1,0 +1,5 @@
+export interface IErrorType {
+	readonly type: string;
+	readonly title: string;
+	readonly defaultMessage: string;
+}