+++ apps/server/src/core/error/interface/feathers-error.interface.ts
@@ -1,0 +1,7 @@
+// import { FeathersError } from '@feathersjs/errors';
+
+export interface FeathersError extends Error {
+	code: number;
+	className: string;
+	type: string;
+}