+++ apps/server/src/modules/task/repo/task.repo.integration.spec.ts
@@ -1,23 +1,17 @@
 import { EntityManager } from '@mikro-orm/mongodb';
-import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { Test, TestingModule } from '@nestjs/testing';
+import { MongoMemoryDatabaseModule } from '@src/modules/database';
 import { TaskRepo } from './task.repo';
-import { MongoMemoryServer } from 'mongodb-memory-server';
 import { Course, Lesson, Submission, Task } from '../entity';
 import { UserInfo } from '../../news/entity';
 
 describe('TaskService', () => {
-	let mongoServer;
 	let module: TestingModule;
 
 	beforeAll(async () => {
-		mongoServer = new MongoMemoryServer();
-		const mongoUri = await mongoServer.getUri();
 		module = await Test.createTestingModule({
 			imports: [
-				MikroOrmModule.forRoot({
-					type: 'mongo',
-					clientUrl: mongoUri,
+				MongoMemoryDatabaseModule.forRoot({
 					entities: [Task, Lesson, Course, Submission, UserInfo],
 				}),
 			],
@@ -26,7 +20,7 @@ describe('TaskService', () => {
 	});
 
 	afterAll(async () => {
-		await mongoServer.stop();
+		await module.close();
 	});
 
 	describe('findAllOpenByStudent', () => {
@@ -35,24 +29,24 @@ describe('TaskService', () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user], color: '#ffffff' });
-				const task = await em.create(Task, { name: 'roll some dice', course, dueDate: new Date() });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user], color: '#ffffff' });
+				const task = em.create(Task, { name: 'roll some dice', course, dueDate: new Date() });
 				await em.persistAndFlush([user, course, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(1);
-				expect(result.data[0]).toHaveProperty('name');
-				expect(result.data[0]).toHaveProperty('duedate');
-				expect(result.data[0]).toHaveProperty('courseName');
-				expect(result.data[0]).toHaveProperty('displayColor');
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(1);
+				expect(result[0]).toHaveProperty('name');
+				expect(result[0]).toHaveProperty('dueDate');
+				expect(result[0].course).toHaveProperty('name');
+				expect(result[0].course).toHaveProperty('color');
 			});
 
 			it('should return a paginated result', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user], color: '#ffffff' });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user], color: '#ffffff' });
 				const tasks = await Promise.all([
 					em.create(Task, { name: 'warmup', course, dueDate: new Date() }),
 					em.create(Task, { name: 'run 100m', course, dueDate: new Date() }),
@@ -60,26 +54,26 @@ describe('TaskService', () => {
 					em.create(Task, { name: '100 pushups', course, dueDate: new Date() }),
 				]);
 				await em.persistAndFlush([user, course, ...tasks]);
-				const result = await service.findAllOpenByStudent(user.id, { limit: 2, skip: 0 });
-				expect(result.data.length).toEqual(2);
-				expect(result.total).toEqual(4);
+				const [result, total] = await service.findAllOpenByStudent(user.id, { limit: 2, skip: 0 });
+				expect(result.length).toEqual(2);
+				expect(total).toEqual(4);
 			});
 
 			it('should be sorted with earlier duedates first', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user], color: '#ffffff' });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user], color: '#ffffff' });
 				const tasks = await Promise.all([
 					em.create(Task, { name: '2nd', course, dueDate: new Date(Date.now() + 500) }),
 					em.create(Task, { name: '1st', course, dueDate: new Date(Date.now() - 500) }),
 				]);
 				await em.persistAndFlush([user, course, ...tasks]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(2);
-				expect(result.data[0].name).toEqual('1st');
-				expect(result.data[1].name).toEqual('2nd');
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(2);
+				expect(result[0].name).toEqual('1st');
+				expect(result[1].name).toEqual('2nd');
 			});
 		});
 
@@ -88,71 +82,71 @@ describe('TaskService', () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user] });
-				const task = await em.create(Task, { name: 'roll some dice', course });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user] });
+				const task = em.create(Task, { name: 'roll some dice', course });
 				await em.persistAndFlush([user, course, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(1);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(1);
 			});
 
 			it('should not return task of other course', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [] });
-				const task = await em.create(Task, { name: 'secret task', course });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [] });
+				const task = em.create(Task, { name: 'secret task', course });
 				await em.persistAndFlush([user, course, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(0);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(0);
 			});
 
 			it('should not return private task of course', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [] });
-				const task = await em.create(Task, { name: 'secret task', course, private: true });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [] });
+				const task = em.create(Task, { name: 'secret task', course, private: true });
 				await em.persistAndFlush([user, course, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(0);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(0);
 			});
 
 			it('should not return task that has a submission by the user', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user] });
-				const task = await em.create(Task, { name: 'roll some dice', course });
-				const submission = await em.create(Submission, { homework: task, student: user });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user] });
+				const task = em.create(Task, { name: 'roll some dice', course });
+				const submission = em.create(Submission, { homework: task, student: user });
 				await em.persistAndFlush([user, course, submission, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(0);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(0);
 			});
 
 			it('should not return task that has a submission by different user', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const otherUser = await em.create(UserInfo, { firstName: 'other', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user] });
-				const task = await em.create(Task, { name: 'roll some dice', course });
-				const submission = await em.create(Submission, { homework: task, student: otherUser });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const otherUser = em.create(UserInfo, { firstName: 'other', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user] });
+				const task = em.create(Task, { name: 'roll some dice', course });
+				const submission = em.create(Submission, { homework: task, student: otherUser });
 				await em.persistAndFlush([user, otherUser, course, submission, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(1);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(1);
 			});
 
 			it('should filter tasks that are more than one week overdue', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user] });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user] });
 
 				const threeWeeksinMilliseconds = 1.814e9;
 				const tasks = await Promise.all([
@@ -164,8 +158,8 @@ describe('TaskService', () => {
 					em.create(Task, { name: 'should have done this recently', course }),
 				]);
 				await em.persistAndFlush([user, course, ...tasks]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(1);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(1);
 			});
 		});
 
@@ -174,64 +168,64 @@ describe('TaskService', () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user] });
-				const lesson = await em.create(Lesson, { course, hidden: false });
-				const task = await em.create(Task, { name: 'roll some dice', lesson, course });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user] });
+				const lesson = em.create(Lesson, { course, hidden: false });
+				const task = em.create(Task, { name: 'roll some dice', lesson, course });
 				await em.persistAndFlush([user, course, lesson, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(1);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(1);
 			});
 
 			it('should return task if lesson is null', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user] });
-				const task = await em.create(Task, { name: 'roll some dice', lesson: null, course });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user] });
+				const task = em.create(Task, { name: 'roll some dice', lesson: null, course });
 				await em.persistAndFlush([user, course, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(1);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(1);
 			});
 
 			it('should not return task in a lesson of a different course', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [] });
-				const lesson = await em.create(Lesson, { course });
-				const task = await em.create(Task, { name: 'roll some dice', lesson, course });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [] });
+				const lesson = em.create(Lesson, { course });
+				const task = em.create(Task, { name: 'roll some dice', lesson, course });
 				await em.persistAndFlush([user, course, lesson, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(0);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(0);
 			});
 
 			it('should not return task in a hidden lesson', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user] });
-				const lesson = await em.create(Lesson, { course });
-				const task = await em.create(Task, { name: 'roll some dice', lesson, course });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user] });
+				const lesson = em.create(Lesson, { course });
+				const task = em.create(Task, { name: 'roll some dice', lesson, course });
 				await em.persistAndFlush([user, course, lesson, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(0);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(0);
 			});
 
 			it('should not return task in a lesson when hidden is null', async () => {
 				const service = module.get<TaskRepo>(TaskRepo);
 				const em = module.get<EntityManager>(EntityManager);
 
-				const user = await em.create(UserInfo, { firstName: 'test', lastName: 'student' });
-				const course = await em.create(Course, { name: 'testCourse', students: [user] });
-				const lesson = await em.create(Lesson, { course });
-				const task = await em.create(Task, { name: 'roll some dice', lesson, course });
+				const user = em.create(UserInfo, { firstName: 'test', lastName: 'student' });
+				const course = em.create(Course, { name: 'testCourse', students: [user] });
+				const lesson = em.create(Lesson, { course });
+				const task = em.create(Task, { name: 'roll some dice', lesson, course });
 				await em.persistAndFlush([user, course, lesson, task]);
-				const result = await service.findAllOpenByStudent(user.id);
-				expect(result.total).toEqual(0);
+				const [result, total] = await service.findAllOpenByStudent(user.id);
+				expect(total).toEqual(0);
 			});
 		});
 	});