+++ apps/server/src/server.controller.spec.ts
@@ -1,22 +1,20 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ServerController } from './server.controller';
-import { ServerService } from './server.service';
 
 describe('ServerController', () => {
-  let serverController: ServerController;
+	let serverController: ServerController;
 
-  beforeEach(async () => {
-    const app: TestingModule = await Test.createTestingModule({
-      controllers: [ServerController],
-      providers: [ServerService],
-    }).compile();
+	beforeEach(async () => {
+		const app: TestingModule = await Test.createTestingModule({
+			controllers: [ServerController],
+		}).compile();
 
-    serverController = app.get<ServerController>(ServerController);
-  });
+		serverController = app.get<ServerController>(ServerController);
+	});
 
-  describe('root', () => {
-    it('should return "Hello World!"', () => {
-      expect(serverController.getHello()).toBe('Hello World!');
-    });
-  });
+	describe('root', () => {
+		it('should return "Schulcloud Server API"', () => {
+			expect(serverController.getHello()).toBe('Schulcloud Server API');
+		});
+	});
 });