+++ apps/server/src/shared/core/repo/index.ts
@@ -1,8 +1,0 @@
-/**
- * Shared repository models, consists of entities and interfaces.
- * Repositories must provide their own interfaces.
- * It is forbidden to reuse existing ones from different layers.
- */
-
-export * from './entity';
-export * from './interface';