+++ apps/server/src/shared/core/repo/interface/index.ts
@@ -1,6 +1,0 @@
-/**
- * Interfaces are used to declare shapes of data which do not need to be instances of a class.
- * Use it for basic functionalities like sort, paginate, filter, select, ... .
- */
-
-export * from './pagination.interface';