+++ apps/server/test/modules/task/task.e2e-spec.ts
@@ -1,12 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ExecutionContext, INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
+import { Request } from 'express';
+import { MikroORM } from '@mikro-orm/core';
 import { ServerModule } from '../../../src/server.module';
 import { JwtAuthGuard } from '../../../src/modules/authentication/guard/jwt-auth.guard';
 import { ICurrentUser } from '../../../src/modules/authentication/interface/jwt-payload';
 
 describe('Task Controller (e2e)', () => {
 	let app: INestApplication;
+	let orm: MikroORM;
 
 	beforeEach(async () => {
 		const moduleFixture: TestingModule = await Test.createTestingModule({
@@ -15,7 +18,7 @@ describe('Task Controller (e2e)', () => {
 			.overrideGuard(JwtAuthGuard)
 			.useValue({
 				canActivate(context: ExecutionContext) {
-					const req = context.switchToHttp().getRequest();
+					const req: Request = context.switchToHttp().getRequest();
 					const user: ICurrentUser = {
 						userId: '0000d224816abba584714c9c',
 						roles: [],
@@ -30,10 +33,16 @@ describe('Task Controller (e2e)', () => {
 
 		app = moduleFixture.createNestApplication();
 		await app.init();
+		orm = app.get<MikroORM>(MikroORM);
+	});
+
+	afterEach(async () => {
+		await app.close();
+		await orm.close();
 	});
 
 	it('/ (FIND)', async () => {
 		const response = await request(app.getHttpServer()).get('/task/dashboard');
-		expect(response.status == 200);
+		expect(response.status === 200);
 	});
 });