+++ src/services/config/publicAppConfigService.js
@@ -16,6 +16,10 @@ const exposedVars = [
 	'FEATURE_ES_COLLECTIONS_ENABLED',
 	'FEATURE_EXTENSIONS_ENABLED',
 	'FEATURE_TEAMS_ENABLED',
+	'I18N__AVAILABLE_LANGUAGES',
+	'I18N__DEFAULT_LANGUAGE',
+	'I18N__DEFAULT_TIMEZONE',
+	'I18N__FALLBACK_LANGUAGE',
 	'JWT_SHOW_TIMEOUT_WARNING_SECONDS',
 	'JWT_TIMEOUT_SECONDS',
 	'NOT_AUTHENTICATED_REDIRECT_URL',
@@ -37,7 +41,9 @@ class PublicAppConfigService {
 	find() {
 		const envs = {};
 		exposedVars.forEach((varName) => {
-			envs[varName] = Configuration.get(varName);
+			if (Configuration.has(varName)) {
+				envs[varName] = Configuration.get(varName);
+			}
 		});
 
 		return Promise.resolve(envs);