+++ CHANGELOG.md
@@ -9,8 +9,47 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ## Unreleased
 
+### Added
+
 - SC-9018; SC-9003 - created schoolsList public endpoint, and jwt secured /schools endpoint
+- SC-9093 - make configured default language and timezone available in config service
+- SC-9092 - delete S3 files after 7 days from user deletion
 - SC-8959 - Add messenger to deletion concept
+
+### Changed 
+
+- remove unneccesary timeout definitions from tests and avoid promise chains
+- SC-6294 Restructure NestJS Sources: Testing, Core Module, Entities, Shared. See details in https://hpi-schul-cloud.github.io/schulcloud-server/
+
+## [26.4.5] - 2021-06-21
+
+### Added 
+
+- SC-9156 - Add maintenance mode for LDAP rewrite
+
+## [26.4.4] - 2021-06-16
+
+### Change
+
+- rename permission TASK_DASHBOARD_VIEW_V3
+
+## [26.4.3] - 2021-06-16
+
+### Change
+
+- SC-9139 - Add a check if user roles should be updated or not to the repo
+
+## [26.4.2] - 2021-06-16
+
+### Fixed
+
+- npm run syncIndex work for not existing collections in the db
+
+## [26.4.1] - 2021-06-15
+
+### Change
+
+- SC-9029 - Change place of the channel creation for RabbitMQ
 
 ## [26.4.0] - 2021-06-11
 
@@ -28,19 +67,19 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 - SC-8898 - parallelize LDAP sync using RabbitMQ
 
-## 26.2.2 - 2021-06-04
+## [26.2.2] - 2021-06-04
 
 ### Fixed
 
 - Fixed dependencies issue
 
-## 26.2.1 - 2021-06-02
+## [26.2.1] - 2021-06-02
 
 ### Added
 
 - SC-9103 - add logging for syncIndexes script
 
-## 26.2.0 - 2021-06-01
+## [26.2.0] - 2021-06-01
 
 ### Added